feat: add twitter support

+++ api/bot.ts
@@ -9,8 +9,8 @@ import { autoRetry } from "@grammyjs/auto-retry";
 import * as Sentry from "@sentry/node";
 import "dotenv/config";
 
-import { getTiktokVideoInfo } from "../src/services";
-import type { TikTokVideoInfo } from "../src/types";
+import { getTiktokVideoInfo, getTweetInfo } from "../src/services";
+import type { TikTokVideoInfo, TweetInfo } from "../src/types";
 
 const { BOT_TOKEN, ENV, SENTRY_DSN } = process.env;
 
@@ -21,6 +21,7 @@ Sentry.init({
 const isDevelopment = ENV === "development";
 
 const TIKTOK_LINK_REGEX = /^https:\/\/(www|vm|vt).tiktok.com\/.*/g;
+const TWITTER_LINK_REGEX = /^https:\/\/(www\.)?(twitter|x).com\/.*/g;
 const SHAWBERTO_REGEX = /Shawberto, you good?/g;
 const DEVBERTO_REGEX = /Devberto, you good?/g;
 
@@ -61,6 +62,70 @@ bot.on("::url").hears(TIKTOK_LINK_REGEX, async (ctx) => {
     await ctx.replyWithVideo(videoUrl.data.play, {
       caption: videoUrl.data.title,
     });
+
+    await bot.api.deleteMessage(ctx.chat.id, loader.message_id);
+  } catch (error) {
+    Sentry.captureException(error);
+
+    await bot.api.editMessageText(
+      ctx.chat.id,
+      loader.message_id,
+      "Error processing link"
+    );
+
+    const timer = setTimeout(() => {
+      void bot.api.deleteMessage(ctx.chat.id, loader.message_id);
+      clearTimeout(timer);
+    }, 1000);
+  }
+});
+
+// Handle Twitter links
+bot.on("::url").hears(TWITTER_LINK_REGEX, async (ctx) => {
+  if (!ctx.message?.text) return;
+
+  const loader = await ctx.reply("Processing link...");
+
+  try {
+    const tweetObject: TweetInfo | undefined = await getTweetInfo(
+      ctx.message.text
+    );
+
+    if (!tweetObject) {
+      // eslint-disable-next-line @typescript-eslint/no-throw-literal
+      throw tweetObject;
+    }
+
+    console.log(JSON.stringify(tweetObject, null, 2));
+
+    let mediaUrl: string | undefined;
+    let replyWith: "video" | undefined;
+
+    if (
+      tweetObject.extended_entities.media[0].type === "video" ||
+      tweetObject.extended_entities.media[0].type === "animated_gif"
+    ) {
+      mediaUrl =
+        tweetObject.extended_entities.media[0].video_info.variants.find(
+          (variant) => variant.content_type === "video/mp4"
+        )?.url;
+
+      replyWith = "video";
+    }
+
+    console.log(mediaUrl);
+
+    if (mediaUrl) {
+      await bot.api.editMessageText(
+        ctx.chat.id,
+        loader.message_id,
+        `Sending ${replyWith}...`
+      );
+
+      await ctx.replyWithVideo(mediaUrl, {
+        caption: tweetObject.text,
+      });
+    }
 
     await bot.api.deleteMessage(ctx.chat.id, loader.message_id);
   } catch (error) {
